Clarified ways of install packages
In the `Installing Packages` section, we mention `Notebook scope` and `Cluster scope`. This is a bit confusing compared to the naming that is used in the Databricks documentation about Notebook-scoped Python libraries. When you run `install.packages("...")` in an R notebook, it actually installs the package on a cluster level and not on a notebook-scoped level (like with `%pip` and `%conda` in Python notebooks).